// frontend-angular/src/app/core/generated-api/model/estadoCita.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Estado actual de la cita medica
 */
export const EstadoCita = {
    Programada: 'Programada',
    Confirmada: 'Confirmada',
    Cancelada: 'Cancelada',
    Completada: 'Completada',
    NoAsistio: 'NoAsistio',
} as const;
export type EstadoCita = typeof EstadoCita[keyof typeof EstadoCita];