// frontend-angular/src/app/core/generated-api/model/historialClinicoResponse.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Una entrada del historial clinico de un paciente
 */
export interface HistorialClinicoResponse { 
    readonly id: number;
    /**
     * Fecha y hora en que se registro la entrada clinica
     */
    fechaHora: string;
    /**
     * Nombre del medico que registro la entrada
     */
    medicoNombre: string;
    diagnostico: string;
    tratamiento?: string;
    /**
     * Notas clinicas adicionales del medico
     */
    notas?: string;
}