// frontend-angular/src/app/core/generated-api/model/horarioAtencionResponse.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConsultorioResponse } from './consultorioResponse';
import { DiaSemana } from './diaSemana';


/**
 * Un rango de disponibilidad horaria asignado a un medico en un consultorio especifico
 */
export interface HorarioAtencionResponse { 
    readonly id: number;
    medicoId: number;
    diaSemana: DiaSemana;
    horaInicio: string;
    horaFin: string;
    consultorio: ConsultorioResponse;
}
export namespace HorarioAtencionResponse {
}